blog-commit-5

+++ database/migrations/2021_10_20_080912_create_blog_posts_table.php
@@ -14,10 +14,10 @@ class CreateBlogPostsTable extends Migration
     public function up()
     {
         Schema::create('blog_posts', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-            $table->integer('category_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('category_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
 
             $table->string('slug')->unique();
             $table->string('title');
@@ -33,8 +33,11 @@ class CreateBlogPostsTable extends Migration
             $table->timestamps();
             $table->softDeletes();
 
-            $table->foreign('user_id')->references('id')->on('users');
-            $table->foreign('category_id')->references('id')
+            $table->foreign('user_id')
+                ->references('id')
+                ->on('users');
+            $table->foreign('category_id')
+                ->references('id')
                 ->on('blog_categories');
             $table->index('is_published');
         });
